added auth usage in question web service

+++ StackExchange/src/java/question/QuestionWebService.java
@@ -3,6 +3,7 @@
  */
 package question;
 
+import Auth.Auth;
 import answer.AnswerWebService;
 import database.Database;
 import java.sql.ResultSet;
@@ -24,7 +25,7 @@ import javax.jws.WebService;
 public class QuestionWebService {
     /* Atribut */
     private final String path = "jdbc:mysql://localhost:3306/stack_exchange";
-    
+    private final String url = "http://localhost:8082/IdentityServices/IdentityChecker";
     /* Method */
     @WebMethod(operationName = "getAllQuestion")
     @WebResult(name="Question")
@@ -84,36 +85,45 @@ public class QuestionWebService {
     @WebMethod(operationName = "addQuestion")
     @WebResult(name="String")
     public String addQuestion(String token, String name, String email, String topic, String content, int userId) {
-        String query = "INSERT INTO question (question_id, asker_name, asker_email, question_topic, question_content, user_id) "
-                + "VALUES (NULL, '" + name + "', '" + email + "', '" + topic + "', '" + content + "', " + userId + ")";
-        Database database = new Database();
-        database.connect(path);
-        database.changeData(query);
-        database.closeDatabase();
-        return "executed";
+        Auth auth = new Auth(token);
+        if(auth.getResponse(url)){
+            String query = "INSERT INTO question (question_id, asker_name, asker_email, question_topic, question_content, user_id) "
+                    + "VALUES (NULL, '" + name + "', '" + email + "', '" + topic + "', '" + content + "', " + userId + ")";
+            Database database = new Database();
+            database.connect(path);
+            database.changeData(query);
+            database.closeDatabase();
+            return "executed";
+        } else return "not executed";
     }
     
     @WebMethod(operationName = "editQuestion")
     @WebResult(name="String")
     public String editQuestion(String token, int id, String topic, String content) {
-        String query = "UPDATE question SET question_topic='" + topic + "', question_content='"
-                + content + "' WHERE question_id = " + id;
-        Database database = new Database();
-        database.connect(path);
-        database.changeData(query);
-        database.closeDatabase();
-        return "executed";
+        Auth auth = new Auth(token);
+        if(auth.getResponse(url)){
+            String query = "UPDATE question SET question_topic='" + topic + "', question_content='"
+                    + content + "' WHERE question_id = " + id;
+            Database database = new Database();
+            database.connect(path);
+            database.changeData(query);
+            database.closeDatabase();
+            return "executed";
+        } else return "not executed";
     }
     
     @WebMethod(operationName = "deleteQuestion")
     @WebResult(name="String")
     public String deleteQuestion(String token, int id) {
-        String query = "DELETE FROM question WHERE question_id = " + id;
-        Database database = new Database();
-        database.connect(path);
-        database.changeData(query);
-        database.closeDatabase();
-        return "executed";
+        Auth auth = new Auth(token);
+        if(auth.getResponse(url)){
+            String query = "DELETE FROM question WHERE question_id = " + id;
+            Database database = new Database();
+            database.connect(path);
+            database.changeData(query);
+            database.closeDatabase();
+            return "executed";
+        } else return "not executed";
     }
     
 }
